Add the quote and footer section of the project
I finished the page. Added some polishing touches.
It would still need some care to make it look exactly
like the assignment but I learned a lot and I think
that the other styling things are just a matter of
personal preference. What matters is that the layout
is pretty much the same as was required.